+++ crud/common/utils.lua
@@ -21,11 +21,12 @@ local GetSpaceError = errors.new_class('GetSpaceError')
 local GetSpaceFormatError = errors.new_class('GetSpaceFormatError', {capture_stack = false})
 local FilterFieldsError = errors.new_class('FilterFieldsError', {capture_stack = false})
 local NotInitializedError = errors.new_class('NotInitialized')
-local StorageInfoError = errors.new_class('StorageInfoError')
 local VshardRouterError = errors.new_class('VshardRouterError', {capture_stack = false})
 local UtilsInternalError = errors.new_class('UtilsInternalError', {capture_stack = false})
 
 local utils = {}
+
+utils.STORAGE_NAMESPACE = '_crud'
 
 --- Returns a full call string for a storage function name.
 --
@@ -35,10 +36,8 @@ local utils = {}
 function utils.get_storage_call(name)
     dev_checks('string')
 
-    return '_crud.' .. name
+    return ('%s.%s'):format(utils.STORAGE_NAMESPACE, name)
 end
-
-local CRUD_STORAGE_INFO_FUNC_NAME = utils.get_storage_call('storage_info_on_storage')
 
 local space_format_cache = setmetatable({}, {__mode = 'k'})
 
@@ -684,13 +683,22 @@ local function determine_enabled_features()
 
     enabled_tarantool_features.netbox_skip_header_option = is_version_ge(major, minor, patch, suffix,
                                                                          2, 2, 0, nil)
+
+    -- https://github.com/tarantool/tarantool/commit/11f2d999a92e45ee41b8c8d0014d8a09290fef7b
+    enabled_tarantool_features.box_watch = is_version_ge(major, minor, patch, suffix,
+                                                         2, 10, 0, 'beta2')
+
+    enabled_tarantool_features.tarantool_3 = is_version_ge(major, minor, patch, suffix,
+                                                           3, 0, 0, nil)
 end
 
 determine_enabled_features()
 
 for feature_name, feature_enabled in pairs(enabled_tarantool_features) do
     local util_name
-    if feature_name == 'builtin_merger' then
+    if feature_name == 'tarantool_3' then
+        util_name = ('is_%s'):format(feature_name)
+    elseif feature_name == 'builtin_merger' then
         util_name = ('tarantool_has_%s'):format(feature_name)
     else
         util_name = ('tarantool_supports_%s'):format(feature_name)
@@ -1120,124 +1128,6 @@ function utils.list_slice(list, start_index, end_index)
     end
 
     return slice
-end
-
---- Polls replicas for storage state
---
--- @function storage_info
---
--- @tparam ?number opts.timeout
---  Function call timeout
---
--- @tparam ?string|table opts.vshard_router
---  Cartridge vshard group name or vshard router instance.
---
--- @return a table of storage states by replica id.
-function utils.storage_info(opts)
-    opts = opts or {}
-
-    local vshard_router, err = utils.get_vshard_router_instance(opts.vshard_router)
-    if err ~= nil then
-        return nil, StorageInfoError:new(err)
-    end
-
-    local replicasets, err = vshard_router:routeall()
-    if replicasets == nil then
-        return nil, StorageInfoError:new("Failed to get router replicasets: %s", err.err)
-    end
-
-    local futures_by_replicas = {}
-    local replica_state_by_id = {}
-    local async_opts = {is_async = true}
-    local timeout = opts.timeout or const.DEFAULT_VSHARD_CALL_TIMEOUT
-
-    for _, replicaset in pairs(replicasets) do
-        for replica_id, replica in pairs(replicaset.replicas) do
-            local master = utils.get_replicaset_master(replicaset, {cached = false})
-
-            replica_state_by_id[replica_id] = {
-                status = "error",
-                is_master = master == replica
-            }
-
-            local ok, res = pcall(replica.conn.call, replica.conn, CRUD_STORAGE_INFO_FUNC_NAME,
-                                  {}, async_opts)
-            if ok then
-                futures_by_replicas[replica_id] = res
-            else
-                local err_msg = string.format("Error getting storage info for %s", replica_id)
-                if res ~= nil then
-                    log.error("%s: %s", err_msg, res)
-                    replica_state_by_id[replica_id].message = tostring(res)
-                else
-                    log.error(err_msg)
-                    replica_state_by_id[replica_id].message = err_msg
-                end
-            end
-        end
-    end
-
-    local deadline = fiber.clock() + timeout
-    for replica_id, future in pairs(futures_by_replicas) do
-        local wait_timeout = deadline - fiber.clock()
-        if wait_timeout < 0 then
-            wait_timeout = 0
-        end
-
-        local result, err = future:wait_result(wait_timeout)
-        if result == nil then
-            future:discard()
-            local err_msg = string.format("Error getting storage info for %s", replica_id)
-            if err ~= nil then
-                if err.type == 'ClientError' and err.code == box.error.NO_SUCH_PROC then
-                    replica_state_by_id[replica_id].status = "uninitialized"
-                else
-                    log.error("%s: %s", err_msg, err)
-                    replica_state_by_id[replica_id].message = tostring(err)
-                end
-            else
-                log.error(err_msg)
-                replica_state_by_id[replica_id].message = err_msg
-            end
-        else
-            replica_state_by_id[replica_id].status = result[1].status or "uninitialized"
-        end
-    end
-
-    return replica_state_by_id
-end
-
---- Storage status information.
---
--- @function storage_info_on_storage
---
--- @return a table with storage status.
-function utils.storage_info_on_storage()
-    return {status = "running"}
-end
-
---- Initializes a storage function by its name.
---
---  It adds the function into the global scope by its name and required
---  access to a vshard storage user.
---
---  @function init_storage_call
---
---  @param string name of a user or nil if there is no need to setup access.
---  @param string name a name of the function.
---  @param function func the function.
---
---  @return nil
-function utils.init_storage_call(user, name, func)
-    dev_checks('?string', 'string', 'function')
-
-    rawset(_G['_crud'], name, func)
-
-    if user ~= nil then
-        name = utils.get_storage_call(name)
-        box.schema.func.create(name, {setuid = true, if_not_exists = true})
-        box.schema.user.grant(user, 'execute', 'function', name, {if_not_exists=true})
-    end
 end
 
 local expected_vshard_api = {