+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+* Asynchronous bootstrap support for storages (#412).
+
 ### Changed
 * Explicitly forbid datetime interval conditions (#373).
+* Storage initialization is now asynchronous by default for Tarantool 3.0+ (#412).
 
 ### Fixed
 * Working with datetime conditions in case of non-indexed fields or